requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
numpy==1.24.3
mysql-connector-python==8.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
PyJWT==2.8.0
requests==2.31.0
